+++ NEWS.md
@@ -1,3 +1,9 @@
+# easyRasch2 0.5.4
+
+- New functions for handling missing data with conditional item infit MSQ by multiple imputation using package `mice`.
+  - **`RMinfitcutoff_mi()`** uses the `mids` object containing multiple imputated datasets (output by `mice::mice()`) to run simulations on each datasets and combines the results
+  - **`RMiteminfit_mi()`** calculates and pools conditional infit from the imputated datasets and optionally uses the `RMinfitcutoff_mi()` output for cutoff values.
+
 # easyRasch2 0.5.3
 
 * New functions **`RMpartgamLD()`** and **`RMpgLDcutoff()`** for evaluating local dependency of item pairs in both directions.